extracted user routes to seperate files, and added authorisation
